Implement event listeners.
